Build features, model for each set individually
Major improvement! Jumped 28 spots in the leaderboard. Not sure if this
is because training separately is better or because of the value shift
bug in my last submission. I'll retry that one after correcting
tomorrow.

I also cleaned up my tokenizer and reduced the amount of words to use as
features. Colors are back in as a predictor, but they seem to do very
little.

I added the proper error calculation to my CV check: Quadratic Weighted
Kappa. CV is optimistic by about .1 against public leaderboard.